backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-jwt-extended==4.5.2
pymongo==4.4.1
python-dotenv==1.0.0
werkzeug==2.3.6
requests==2.31.0  # <-- Ajoutez cette ligne